CA-174475 Make 2048-bit Diffie-Hellman prime for TLS certificate
When creating a certificate to use for TLS, we were generating 512-bit
Diffie-Hellman parameters. This is rejected as too short by several
clients. Now we have changed from 512 to 2048.

Generating DH parameters of this length tends to take anything from a
few seconds to a few minutes.

This commit does not address the issue of existing certificates from
XenServers upgraded from (say) XenServer 6.5.

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>