rrdd_shared: Remove localhost_uuid global
localhost_uuid was initialized by querying the inventory.
Doing this raised an exception on hosts which did not have
the inventory file in the right place, making unit testing
difficult.   The Inventory module caches the contents of
the inventory file in a hashtable, so looking the UUID up
directly should not be much slower than getting it from a
global.

Signed-off-by: Euan Harris <euan.harris@citrix.com>