CP-13189 TLS ciphersuite TLS_RSA_WITH_AES_128_CBC_SHA
For outbound connections from XAPI, this commit implements the
recommendation from the Citrix security team that XenServer should
use only ciphersuite TLS_RSA_WITH_AES_128_CBC_SHA for TLSv1.2.

(The exception is when the host ssl-legacy setting is switched on for
backward compatibility).

This recommendation is based on a combination of criteria:
* technical (security) considerations
* legal considerations
* compatibility with third-party TLSv1.2 clients

This commit relies on a corresponding commit in another repository to
set the ciphersuite accepted for incoming connections to xapi, so that
XenServer hosts will be able to communicate with one another.

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>