Async: use `Monitor.try_with ~extract_exn:true` so we can match on the exn
We want to retry `Socket.connect` if the Unix domain socket is missing or
returning ECONNREFUSED as these indicate either (i) the server has never
started; or (ii) the server is temporarily offline. We want to make other
failures (such as EACCES) hard failures and quit.

If we don't provide `extract_exn:true` then the exception is wrapped with
additional information (including backtraces) which means we can't pattern
match as easily.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>