In preconfigure_host, put cluster stack in local DB and use for scripts
We have to store the cluster stack-in-use in the local DB as well to avoid the
chicken-and-egg problem on start up.

Added a "cluster-stack-root" xapi.conf option. This is a path to a directory of
cluster stack implementations, each having its own directory within the root.
We also add "cluster-stack-default", which will be "xhad" for the foreseeable
future. The xhad scripts and tools will be moved into the xhad dir inside the
cluster-stack-root.

Finally, we make sure that we call the HA scripts in the cluster-stack-root
with the right PATH.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>