CA-178692: Only kill xenvmd(s) when becoming a slave
It's possible that `become_another_masters_slave` can also be executed on
a host that is already a slave. In this case the current host

  a. Doesn't need to kill any xenvmd instances since there are none; and
  b. Can't because this involves talking to the host it thinks is the master.

Consider the three host example: A, B and C where A is the pool master.

  1. Host A dies;
  2. Pool.emergency_transition_to_master is called on host B;
  3. Pool.recover_slaves is called on host B.

In this scenario, host B calls Pool.emergency_reset_master to host C. Host C is
already a slave and is remaining a slave but acknowledging B as the new master.
It therefore cannot make use of any functions that required (what it thinks is)
the master (host A) since these will hang.

Worse still, if host A were to come alive again, these calls would unblock when
host A starts up as mater and set the master of C to B (regardless of what's
happened in between).

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>