CA-178240: Shrink SR.scan race window
The longer the time between calling Storage_access.SR.scan and checking
the database to see what needs updating, the more time there is for
something to change behind the scenes - for instance, the SM GC could
call VDI.forget. Moving the SR.stat call means that this window is much
smaller.

Signed-off-by: John Else <john.else@citrix.com>