CP-12902: Expose the control interface over a message queue
If the message switch is enabled, then this will serve control RPCs
over the well-known queue name (org.xen.xcp.rrd). If the message
switch is not enabled, then there is no change.

Note that we are still using a mix of Xcp_service code and custom
Unix domain socket code because Xcp_service is not able to create the
HTTP server configuration that we need. This means we must take care
to only call `Xcp_service.make` when the message switch is enabled,
or else it will erase the Unix domain socket that we just created.

This also changes the command-line syntax to be more consistent with
other services. In particular the init.d script must use

  --pidfile <foo> --daemon true

Signed-off-by: David Scott <dave.scott@citrix.com>