Obtain cluster stack constraints and use them
Constraints come from the storage backend as well as the cluster stack
implementation.  A storage backend may express that it requires a certain
cluster stack, and may specify a list of alternatives. Likewise, a cluster
stack may specify a number of SR types that it is able to use to heartbeating
and state.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>