Add the concept of XenAPI "extensions"
An "extension" is a XenAPI call which is implemented by an external
entity, currently limited to a program in a pre-configured directory
(defaults to /etc/xapi.d/extensions). After an incoming API is
authenticated, the program is executed with the argument "--xmlrpc"
and the full XenAPI RPC request is send on stdin as XMLRPC. The program
should print the XenAPI RPC response on stdout.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>