Allow undeclared XenAPI extensions to be added at runtime
If we cannot find the definition of an API, then we check to see
whether a corresponding extension can be found. Callers can check
whether a server has the extension first via the `pool.has_extension`
API.

Note we call the extension using the same RBAC settings as the
`pool.has_extension` call i.e. POOL_ADMIN.

Note that extensions are "pool-level": this means that we execute
them on the pool master. If an implementation needs to contact
other hosts behind the scenes, this is fine. Unlike plugins, users
cannot ask for the code to run on a particular host.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>