Add dbsync-exception list to the config file
A failing dbsync causes xapi not to start, which makes the system
unmanageable. Before this patch the only way to skip part of the
dbsync wa to add a pool other-config key; however this is not much
good if xapi won't start and the database is offline.

After this patch you can add

  disable-dbsync-for=sync_pgpus sync_foo sync_bar

in /etc/xpai.conf to fix this problem.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>