Refactor module Storage_migrate.State
Instead of allowing the active_send, active_recv and active_copy tables
to each to contain send, receive and copy operations, restrict each
table to only contain its respective operation types.

Eliminate the opty type, and instead use a GADT to allow operations to
be added the correct table without having to specify which table that
is.

Signed-off-by: John Else <john.else@citrix.com>