CAR-1744: change result type from unit to Rpc.t
So far all xenopsd operations do not have return value. However, in some cases,
we do need to return data/information back to the caller, hence the unit type
no longer fits.

In theory, different operations might want to return different types of data
and the possibility is unlimited. Here we use Rpc.t as a generic representation
of all the possible return types (also be consistent with the type argument of
the "Failed" variant here). It's up to the caller to eventually decode the
value properly. As the cases of requiring a return value are still rare, it
won't be a hassle.

Signed-off-by: Zheng Li <dev@zheng.li>