CAR-1744: syntax extension for vm plugin call --- args:key:file=filename
In the original design of CAR-1744, the CLI param "args:script" represents a
script as a single string to be passed from the CLI client (xe) to CLI server
(xapi CLI backend). As scripts usually carry special characters which could
mangle the body of a CLI request (via http), we added string escaping on the
client side and unescaping on the server side. This worked as expected by
itself. However some testcases revealed an issue here (CA-182869) : when a
older version of xe client talks to a newer version of server or when a new
version of xe client talks to a older version of server, the escape/unescape
layers don't match hence will fail. We believe that having a single xe client
to talk to multiple servers of different releases/versions is a valid use case,
so this is a bug we need to fix.

The first attempt was to make use of the CLI protocol version number that exist
on both CLI client side and server side.  Theoretically, the server side could
then handle the string unescaping differently based on the CLI protocol number
of the client side, and vice versa. However there is some legacy baggage in the
way xe client/server interact, in that the server side will only send out its
CLI protocol info to the client side _after_ the client has sent the full
request. Therefore, the client has no way to do string escaping differently
based on the server's CLI protocol version (which only comes later). This makes
it infeasible to get a clean solution.

Jon Ludlam suggested a syntax extension to the vm-call-plugin CLI syntax, with
which a user can pass the content of a file (with any special characters in it)
as the "value" argument by specifying a the file's name (i.e. local path). The
syntax is generic, for any arguments in the form "args:key=value", the
alternative syntax "args:key:file=filename" will instead pass the content of
"filename" in place of the "value" argument, hence semantically equivalent to
"args:key=content_of_filename". As the file content will be effectively loaded
and unstreamed by the CLI frontend via a special "Load" primitive in our
implementation, we no longer need to worry about having special charters
mangling the CLI command line or CLI http request and is also backward
compatible.

Signed-off-by: Zheng Li <dev@zheng.li>