XOP-681: Fix some concurrency issue on host patching
The current patch applying implementation has some concurrency issue. During
the process of applying a pool patch onto a host (which could last long
depending on the patch size):

  * We allow to launch another instance of the same job and cause them
    conflict with each other

  * We allow to destroy the pool patch record so that when the patch applying
    process finishes it will encounter various DB errors

This patch reuses the host_patch.applied field (which doesn't seem to be in
use as it always equals true) to record the patch applying status. When
host_patch.applied=false, it implies the patching is still in progress; and
when host_patch.applied=true, it implies the patch has been applied
successfully (like before). If the patch is not on a host at all, there won't
be the host_patch object existing (like before).

Various logic in xapi_pool_patch.ml has been update to reflect this.

Signed-off-by: Zheng Li <dev@zheng.li>