CA-184513: Fix deadlock with PBD.plug and stats VDI creation
Creating a VDI is not possible while the SR is marked for plug.

We should therefore do this in the message forwarding layer once the SR has
been unmarked as being in the process of being plugged.

For symmetry we also move the archiving of SR RRDs on PBD.unplug to the message
forwarding layer.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>