CA-184513: i_am_srmaster only true for local non-shared SRs
The previous incarnation of this function assumed you would only call this for
SRs for which there was a PBD plugged on localhost. This turned out to not be
the case and was returning true on the pool master for all local SRs on the
slaves.

Assume nothing. Do the right thing(TM).

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>