CA-178504: postpone Db access until network brings up
With the recent changes on SSL, there are two instances of Db accessing added
to the first phase of server_init and before the network set up. One is the
"set_stunnel_legacy_db", the other is Xapi_mgmt_iface.restart_stunnel (called
by bring_up_management_if). If, at that point, the DB is not accessible (say
the master is down or can't be reached from the network), it will block the
server_init process and stop xapi from listening on the SSL port properly.

This patch tries to solve this by avoiding DB access in the first phase of
server_init and reconfigure the SSL legacy param afterwards (when the DB access
can succeed).

Signed-off-by: Zheng Li <dev@zheng.li>