Update Initialize-Environment.ps1
Changed to prevent the init procedure from crashing in a "initializationscript" on Start-Job in Powershell. The InitializationScript seems to ignore the "errorAction" setting and terminates the job when the Variable XenServer_Environment_Initialized does not exist. 
I found this way of checking for the variable on stackoverflow (http://stackoverflow.com/questions/3159949/in-powershell-how-do-i-test-whether-or-not-a-specific-variable-exists-in-global). Neat thing on this is that it will not generate a exception when the variable is not set so it should work in an InitializationScript environment as well