Add upgrade function for CPU feature strings
Upgrades a VM's feature set based on the host's one, if needed.  The output
will be a feature set that is the same length as the host's set, with a prefix
equal to the VM's set, and extended where needed.  If the current VM set has 4
words, then we assume it was last running on a host that did not support
"feature levelling v2". In that case, we cannot be certain about which host
features it was using, so we'll extend the set with all current host features.
Otherwise we'll zero-extend.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>