CA-186162: Only do SR book-keeping on master PBD plug/unplug
We should only call SR.scan (and associated callbacks) when we plug the PBD on
the master. Before this patch we were calling these functions on every PBD
plug. The code would run on the master, since it was in the message forwarding
layer. Since commit 5b6d293, there is now a constraint that the SR must be
plugged to the master first and unplugged last.

This means that we now only process the DR metadata VDIs, process the SR level
stats when un/plugging the master PBD.

We also now only call SR.update on when the master PBD is plugged.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>