CA-182896: extra SXM precheck for CD missing on the receiver side
In the existing VM metadata import code (used by SXM), if the request was for
SXM, we basically exempt the whole operation from CD existence checking, with
the consideration that we're going to eject any CD before the actual migration.

It's not right to do this any more. The migration VM might be in a suspended
status itself (hence a VBD eject is disallowed on the sender side), and also
the VM might carry checkpoints (which are basically suspended VMs).

On the receiver side, our import code has been rather tolerant regarding
missing CDs (by auto ejecting them during the import if we could), but still we
couldn't handle the case where a running/suspended PV guest attached to a CD
that is missing in the environment.

This change adds the extra precheck in place. It still try to be as tolerant as
we could --- only prevent cases we can't handle on the receiver side and also
take into consideration that sender could eject the CD (unless the VM is
suspended). It will raise things up if a SXM is not possible due to this
reason, which will reflect via VM.assert_can_migrate in the cases of SXM.

Signed-off-by: Zheng Li <dev@zheng.li>