Support SXM from pre-levelling-v2 hosts
To support VMs migrated from hosts that do not support the new CPU levelling
("levelling v2"), the `host.cpu_info:features` key is set to (approximately)
what it would be on such hosts.  To do this, we take the HVM features of the
host, and add in any features that we don't care about for migration, such that
the absence of such features will not cause a migration to be disallowed from
the sender.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>