CA-185923: Use a sane API for archiving/pushing SR RRDs
The interface before this patch involved asking an RRD server where for the
path that it would store the archive should it succeed, then an archive call,
and then checking that the file was there since in the error case there are no
exceptions to return.

Similarly for restoring from an archive a client would have to ask the server
where it expects these archives to be to restore, put the file there and then
call restore.

This patch simplifies the interface such that:

* The restore function takes a path to restore from in addition to the uuid;
* The archive function returns a path to the archive;
* An exception is defined so that we don't have to silently fail.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>