CA-184300: Add an "extras" param for VM.build/restore
To workaround an Xen regression bug, we are requested to pass an extra argument
to the xenguest CLI with some addtional PCI information.

Instead of adding a param specifically for PCI, here we opt to have a more
general param named "extras" for these ad-hoc cases. Using this param, we just
pass a string list down the calling stack until it reaches the bottom. For
the time being, these strings will then be attached to the end of the xenguest
CLI. However, the param itself is rather neutral and can be interpreted at
different layers, so it can also be used for other purpose (on demand) with
minimum modification.

Along with the change, we also did very slight refactoring on the labelling and
ordering of some params to make them more consistent. None of these is externally
visible on the API layer though.

Signed-off-by: Zheng Li <dev@zheng.li>