CA-187017: VM.get_state: don't try to recurse into non-directories in xenstore
Work around strange behaviour in the OCaml xenstore client library so that
VM.get_state doesn't try to read invalid paths from xenstore, each of which is
guaranteed to result in EINVAL from xenstored. This behaviour causes unnecessary
load on xenstore.

The behaviour is as follows. Consider the following xenstore tree:

  /local/domain/1/dir0 = ""
  /local/domain/1/dir0/key0 = "a"
  /local/domain/1/dir0/key1 = "b"
  /local/domain/1/key1 = "c"

Here's the behaviour of two xenstore commands:

  1. xs.Xs.directory "/local/domain/1/dir0"
      -> ["key0"; "key1"]

  2. xs.Xs.directory "/local/domain/1/key1"
      -> [""]

Perhaps more appropriate behaviour would be to return [] in the second example.

But the existing [""] behaviour means that xenopsd's ls_lR function thinks that
there's a sub-directory called "", so tries to recurse into
("/local/domain/1/dir0" ^ "/" ^ "") i.e. "/local/domain/1/dir0/" which is
invalid.

This patch squashes any reported entry that matches the empty string.

Signed-off-by: Jonathan Davies <jonathan.davies@citrix.com>