CP-14811: Support SXM from pre-levelling-v2 hosts
When a pre-levelling-v2 host is asked to storage-migrate a VM to
a post-levelling-v2 host, the older host will compare its local CPU
features to the 'features' flag sent by the new host.   The new host
must send its features key in the old, pre-levelling-v2 format so that
the older host can understand it.

Signed-off-by: Euan Harris <euan.harris@citrix.com>