Correct the use of Xenctrl.Halt
Xenctrl.Halt has always been incorrect in the libxc bindings (and should have
been Watchdog).  This was fixed upstream in
http://xenbits.xen.org/gitweb/?p=xen.git;a=commitdiff;h=87a16694248e80195943741bb9f857edd6a6bf51,
but subsequently causes a compilation error in xenopsd.

Instead of shutting the domain down with a Watchdog signal, use Poweroff.
This fixes the build against upstream Xen.  (The "halt" case in
xenops_server_xen.ml was just a debug case, so remove it completely.)

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>