CP-14790: Restrict copy/clone of Windows-update VMs
Do the Windows-update-feature licence check in Xapi_vm_clone.clone
before cloning/copying the discs, rather than in copy_vm_record
which is called afterwards.  This is safe because copy_vm_record is
not called from anywhere else and is not exposed in the .mli file,
and Xapi_vm_clone.clone is common code called by VM clone, copy,
snapshot and checkpoint.

We allow creation of snapshots and checkpoints without checking the
licence.

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>