CA-189771: If host does not have new feature keys, don't update VM features
During migration, the master tries to update the VM's CPU features to
the new levelling v2 format.   This is the correct thing to do when the
VM is moving from a non-upgraded host to an upgraded host during rolling
pool upgrade.  However, during RPU the master may also move a VM from
one non-upgraded host to another non-upgraded host.   In this case the
flags should not be updated.   In fact, they cannot be updated, because
the hosts do not have the levelling v2 features_pv and features_hvm flags.
Their absence causes update_cpu_flags to fail with a 'Not_found' exception
which bubbles up from Cpuid_helpers.get_cpu_flags.

This patch makes update_cpu_flags a no-op when the host does not have the
required features.   This should only occur during a 'sideways' migration
where both hosts are the same version and have the same flags:

  * If the source host is newer than the destination, the 'downhill'
    migration will not be permitted.   This means that we don't have
    to do anything special when the VM has new-style features and the
    host has old-style  features.
  * If the source host is older than the destination, the 'uphill'
    migration will be permitted.   update_cpu_features will upgrade
    the VM's old-style features to match the host's new-style
    features.
  * If the source host is the same version as the destination, the
    'sideways' migration will be permitted, but no upgrade will be
    needed because there is no change of version.

Signed-off-by: Euan Harris <euan.harris@citrix.com>