CP-15105: VM.has_vendor_device
A new VM field intended to replace auto_update_drivers.

This commit is the first stage, adding the new field so that new
code can refer to it immediately and existing code can be changed to
refer to it, and then it will be possible to remove the old field.

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>