CP-14791: vm-install: apply vendor-device recommendation
In the implementation of the xe vm-install command, just before the
call to VM.provision, we attempt to set VM.has_vendor_device to the
value from the VM.recommendations field (defaulting to false if not
specified in the recommendations xml).

If the licence feature-set does not permit setting has_vendor_device,
we write a warning to stderr and continue the installation without
setting that flag.

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>