CA-192191: Upgrade CPU flags after domain has migrated
Previously, we would upgrade a VM's CPU flags before asking the host on
which it was resident to migrate it.  If the older host tried to read
the CPU flags while preparing to migrate the VM, it would see new-style
feature flags, which it could not understand, causing it to raise an
exception.

Now, we upgrade the CPU flags after the VM has been migrated.   During an
upgrade, if the VM does not have appropriate feature flags, xenguest
will give it maximum features for the destination host, then Xapi will
apply proper new-style features which will be used for all subsequent
migrations.   The older host will never see the new-style features.

Signed-off-by: Euan Harris <euan.harris@citrix.com>