CP-14790: No licence check for VM.revert and import
Even when creation of Windows-Update VMs is not allowed by the
licence, we want to allow it when it is done in the course of an
import or revert-to-snapshot.

In both cases, we avoid the check for the Windows-Update vendor-device
licence feature by using the relevant helper function instead of
calling Client.VM.create_from_record directly.

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>