CA-182919: Honour the plugin-default config option
This config option is supposed to enable all datasources regardless of whether
the plugin exports them as default or not. It is the option that is set by the
xe-enable-all-plugin-metrics script. However, the ref that it set (in some
shared library) was never used to determine whether the datasource would be
enabled or not. This ref will probably be removed since it is vestigial and we
are storing this setting local to rrdd since it is an config option for this
program.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>