CA-182919: Remove ds_default global mutable
If a user of this library wants to enable non-default plugins they can do so.
We should not have a mutable variable in this library that they set that tweaks
the datasource's default value. The default value should remain the same even
if the user wants to ignore that and export it anyway.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>