CA-173605: avoid memory double counting
In the fix for CA-155888, we solved a concurrency issue. It was due to the fact
that the memory transfer from the squeezed to the domain wasn't atomic, so that
the same memory might be allocated to two competing domains. However, the method
we used was quite conservative: the memory reservation for a domain will only
be dismissed _after_ the completion of the whole domain build process. It means
that, at certain stage during the this process, when we calculate the used/free
memory on the host, the same amount of memory could be counted twice: once as
the memory physically allocated to an established domain, once as the memory
reservation (before it is eventually dismissed). This can be observed more
clearly in cases of domain migration, where the whole process takes rather long
time. The consequence of this is, under certain circumstances, a memory
allocation involved domain operation (like VM.start) that should succeed might
be failed by squeezed, as it considers the host having less free memory (or,
more occupied memory).

This patch doesn't fundamentally change how the CA-155888 fix works. Instead,
it adds some clue information to the memory reservation record, so that when
counting the total memory reservation on a host, the calculation can use that
information to figure out that a domain might have already been built with that
memory allocation so that the reservation shouldn't be counted again here.

Signed-off-by: Zheng Li <dev@zheng.li>