CA-195644: Handle the change in the xs-firstboot package
XenServer has moved the firstboot service to systemd and this does not support
arbitrary actions. As such, when we eject from a pool, calling `service
firstboot activate` does not work.

The package has been revised and we now need to call the script directly
(rather than exercise via `service`) and the verb has been changed to be more
semantically meaningful: from `activate` to `reset-and-reboot`.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>