CP-15277: Read feature/hotplug/vbd|vif from xenstore
Populates vm_guest_metrics.can_use_hotplug_vbd based on the value of
xenstore node /local/domain/N/feature/hotplug/vbd which may be written
by the guest in domain N:
"0" -> `no
"1" -> `yes
Anything else, or absent, -> `unspecified

Does the equivalent for vdi as well as vbd.

This is as specified in docs/misc/xenstore-paths.markdown in the
xen.git repository at http://xenbits.xen.org which specifies:
[start-quote]
A domain may write information about installed PV drivers using
paths of this form.

    ~/feature/hotplug/vif = ("0"|"1") [w]
    ~/feature/hotplug/vbd = ("0"|"1") [w]

By setting these paths to "1" a guest can indicate to a toolstack
that it is capable of responding immediately to instantiation of,
respectively, new vif by bringing online a new PV network device or
a new vbd by bringing online a new PV block device.
If the guest sets this path to "0" then it is indicating that it is
definitely unable to respond immediately and hence the toolstack should
defer instantiaton to the next VM start. However, if the path is absent
then the toolstack may attempt the operation.
[end-quote]

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>