CA-195652: Perform cleanup on all error paths of VM.pool_migrate
Make sure that the laundering of exceptions into API errors doesn't forgo the
cleanup and synchronisation required in Xapi on these specific error paths.

Also, this patch removes the duplicate pool_migrate_complete API call. This
was called on completion of the migration on both the sending host and the
receiving host, both using the API and specifying the same host. This means
that the same logic will be executed twice.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>