CA-195652: Fix VM power-state syncing with Xenopsd on Xapi restart
If the VM metadata existed in Xenopsd then Xapi would set the state to running
and resident-on to localhost regardless of the power-state of the VM in
Xenopsd. It's possible that this VM metadata may still exist in Xenopsd so we
should check the power-state according to Xenopsd and update Xapi (and Xenopsd)
accordingly.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>