CA-195652: Cleanup new domain on VM_receive_memory failure
Before this patch there was a hole in the exception handling where the
newly-created destination domain wouldn't be cleaned up if the VM_restore into
it failed but it would if later operations failed.

This patch unifies the error handling in VM_receive_memory so that it always
cleans up the domain it creates on failure.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>