Use SR.is_tools_sr field in favour of Helpers.is_tools_sr
This is a new read-only field that can now be used to reliably find out whether
an SR is the Tools SR.

The caching done in the Helpers function is not needed
anymore, since the users of it can just get the field from the SR record that
they already have, or add one quick DB-get alongside several other ones.

The helpers can therefore be removed, which simplifies the code.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>