CA-195652: Cause Xenopsd to cleanup before creating new domains
Before this patch there was a window during a failed migration where the
VM's power-state is still running but Xenopsd is destroying the domains†
to cleanup the failed operation. At this point it was possible to submit
a VM.hard_reboot API and have Xenopsd start a new domain for the VM.
However, then the event generated from Xenopsd for the cleanup is
processed by Xapi and the VM's power-state is set to halted. The Xenopsd
event for the reboot is ignored because the VM is marked as not resident
on localhost. At this point there was a domain running for the VM but
the power-state was halted allowing a VM.start API call to cause Xenopsd
to start a second running domain for the same VM.

To solve this, we synchronise the state of the world with Xenopsd before
any operation that would cause a new domain to be created for a given
VM. To do this we call refresh_vm. This injects an event for the VM into
Xenopsd (using UPDATES.refresh_vm) which will cause Xenopsd to do
a VM_check_state and shutdown any old VMs. It also then injects
a barrier into the Xenopsd event queue and synchronises on it so we can
be sure that the Xenopsd on the host has put the host into a consistent
state with respect to the VM.

We also now have to handle the case where the VM state is not as we
thought after syncing with Xenopsd. By this point all checks from Xapi
have passed and the operation has been forwarded to the host on which
the VM was marked as resident on. If after cleanup the state was not as
expected we raise a Bad_power_state. This has the right semantics
because the syncing causes the events for the VM to be flushed and
processed. So the following are sensible:

1. If after cleaning up during a Xapi_xenops.reboot the VM no longer
   exists in Xenopsd then this means that the VM should have been halted
   in Xapi but there were still outstanding (probably async. cleanup)
   operations to be flushed by Xenopsd. So we raise Bad_power_state with
   expected Running and actual Halted; and

2. If after cleaning up during a Xapi_xenops.start the VM already exists
   in Xenopsd before we send the metadata then this means the VM should
   have been running in Xapi but there were still outstanding (probably
   async. cleanup) operations to be flushed by Xenopsd. So we raise
   Bad_power_state with expected Halted and actual Running.

In both of these cases the Xapi_xenops event thread will now have set
the power_state and resident_on fields correctly.

†: There is also a race in Xenopsd during a localhost migration where
the sending and receiving sides of Xenopsd (in different concurrent
threads) can sometimes cleanup the same domain rather than one-each.
This race is also required for the failure mode described above. This
cannot be fixed as there isn't a general remote XenopsAPI for the sender
to use to wait for the receiver. Instead there is just the hand-rolled,
un-versioned migration HTTP API which needs to remain stable for RPU.
This patch makes Xapi/Xenopsd robust to these lingering domains.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>